refactor(blocks): split TokenDetail into composable subcomponents
Move the sections of TokenDetail — header, composite preview, composite
breakdown, alias chain, aliased-by tree, axis variance matrix, usage
snippet — into separate modules under src/token-detail/ and re-export
each as a standalone block. TokenDetail itself keeps the same DOM
output and props; it's now a ~55-line composition.
