+++ packages/blocks/src/index.ts
@@ -8,5 +8,18 @@ export { MotionPreview, type MotionPreviewProps, type MotionSpeed } from '#/Moti
 export { ShadowPreview, type ShadowPreviewProps } from '#/ShadowPreview.tsx';
 export { StrokeStyleSample, type StrokeStyleSampleProps } from '#/StrokeStyleSample.tsx';
 export { TokenDetail, type TokenDetailProps } from '#/TokenDetail.tsx';
+export { AliasChain, type AliasChainProps } from '#/token-detail/AliasChain.tsx';
+export { AliasedBy, type AliasedByProps } from '#/token-detail/AliasedBy.tsx';
+export { AxisVariance, type AxisVarianceProps } from '#/token-detail/AxisVariance.tsx';
+export {
+  CompositeBreakdown,
+  type CompositeBreakdownProps,
+} from '#/token-detail/CompositeBreakdown.tsx';
+export { CompositePreview, type CompositePreviewProps } from '#/token-detail/CompositePreview.tsx';
+export { TokenHeader, type TokenHeaderProps } from '#/token-detail/TokenHeader.tsx';
+export {
+  TokenUsageSnippet,
+  type TokenUsageSnippetProps,
+} from '#/token-detail/TokenUsageSnippet.tsx';
 export { TokenTable, type TokenTableProps } from '#/TokenTable.tsx';
 export { TypographyScale, type TypographyScaleProps } from '#/TypographyScale.tsx';
