+++ packages/blocks/src/token-detail/styles.ts
@@ -1,0 +1,239 @@
+import type { CSSProperties } from 'react';
+
+export const styles = {
+  wrapper: {
+    fontFamily: 'var(--sb-typography-sys-body-font-family, system-ui)',
+    fontSize: 'var(--sb-typography-sys-body-font-size, 14px)',
+    color: 'var(--sb-color-sys-text-default, CanvasText)',
+    background: 'var(--sb-color-sys-surface-default, Canvas)',
+    padding: 16,
+    borderRadius: 6,
+    border: '1px solid var(--sb-color-sys-border-default, rgba(128,128,128,0.2))',
+  } satisfies CSSProperties,
+  heading: {
+    margin: 0,
+    fontFamily: 'ui-monospace, SFMono-Regular, Menlo, monospace',
+    fontSize: 16,
+  } satisfies CSSProperties,
+  subline: {
+    display: 'flex',
+    alignItems: 'center',
+    gap: 8,
+    margin: '4px 0 12px',
+    fontSize: 12,
+    opacity: 0.8,
+  } satisfies CSSProperties,
+  typePill: {
+    display: 'inline-block',
+    padding: '2px 6px',
+    borderRadius: 4,
+    fontSize: 10,
+    letterSpacing: 0.5,
+    textTransform: 'uppercase',
+    background: 'var(--sb-color-sys-surface-muted, rgba(128,128,128,0.15))',
+  } satisfies CSSProperties,
+  description: {
+    margin: '0 0 12px',
+    opacity: 0.85,
+  } satisfies CSSProperties,
+  sectionHeader: {
+    fontFamily: 'ui-monospace, SFMono-Regular, Menlo, monospace',
+    fontSize: 11,
+    textTransform: 'uppercase',
+    letterSpacing: 0.5,
+    opacity: 0.6,
+    margin: '12px 0 6px',
+  } satisfies CSSProperties,
+  chain: {
+    display: 'flex',
+    flexWrap: 'wrap',
+    gap: 6,
+    alignItems: 'center',
+    fontFamily: 'ui-monospace, SFMono-Regular, Menlo, monospace',
+    fontSize: 12,
+  } satisfies CSSProperties,
+  chainNode: {
+    padding: '2px 6px',
+    borderRadius: 4,
+    border: '1px solid var(--sb-color-sys-border-default, rgba(128,128,128,0.2))',
+  } satisfies CSSProperties,
+  arrow: {
+    opacity: 0.5,
+  } satisfies CSSProperties,
+  themeTable: {
+    width: '100%',
+    borderCollapse: 'collapse',
+    tableLayout: 'fixed',
+    fontSize: 12,
+  } satisfies CSSProperties,
+  themeRow: {
+    borderBottom: '1px solid var(--sb-color-sys-border-default, rgba(128,128,128,0.15))',
+  } satisfies CSSProperties,
+  themeCell: {
+    padding: '6px 8px',
+    verticalAlign: 'middle',
+  } satisfies CSSProperties,
+  swatch: {
+    display: 'inline-block',
+    width: 14,
+    height: 14,
+    verticalAlign: 'middle',
+    marginRight: 6,
+    borderRadius: 3,
+    border: '1px solid var(--sb-color-sys-border-default, rgba(0,0,0,0.1))',
+  } satisfies CSSProperties,
+  snippet: {
+    display: 'block',
+    padding: '8px 10px',
+    borderRadius: 4,
+    background: 'var(--sb-color-sys-surface-muted, rgba(128,128,128,0.1))',
+    fontFamily: 'ui-monospace, SFMono-Regular, Menlo, monospace',
+    fontSize: 12,
+    whiteSpace: 'pre',
+    overflow: 'auto',
+  } satisfies CSSProperties,
+  missing: {
+    padding: 12,
+    opacity: 0.7,
+  } satisfies CSSProperties,
+  typographySample: {
+    padding: '8px 0',
+  } satisfies CSSProperties,
+  shadowSample: {
+    width: 140,
+    height: 56,
+    background: 'var(--sb-color-sys-surface-raised, #fff)',
+    border: '1px solid var(--sb-color-sys-border-default, rgba(128,128,128,0.15))',
+    borderRadius: 6,
+  } satisfies CSSProperties,
+  borderSample: {
+    width: 140,
+    height: 56,
+    background: 'var(--sb-color-sys-surface-raised, transparent)',
+    borderRadius: 6,
+  } satisfies CSSProperties,
+  gradientSample: {
+    width: 220,
+    height: 56,
+    borderRadius: 6,
+    border: '1px solid var(--sb-color-sys-border-default, rgba(128,128,128,0.15))',
+  } satisfies CSSProperties,
+  strokeStyleLine: {
+    height: 0,
+    borderTopWidth: 4,
+    borderTopColor: 'var(--sb-color-sys-text-default, CanvasText)',
+    width: 220,
+  } satisfies CSSProperties,
+  strokeStyleSvg: {
+    width: 220,
+    height: 24,
+    color: 'var(--sb-color-sys-text-default, CanvasText)',
+  } satisfies CSSProperties,
+  strokeStyleFallback: {
+    fontFamily: 'ui-monospace, SFMono-Regular, Menlo, monospace',
+    fontSize: 12,
+    color: 'var(--sb-color-sys-text-muted, CanvasText)',
+  } satisfies CSSProperties,
+  colorSwatchRow: {
+    display: 'flex',
+    gap: 1,
+    borderRadius: 6,
+    overflow: 'hidden',
+    border: '1px solid var(--sb-color-sys-border-default, rgba(128,128,128,0.2))',
+    width: 220,
+    height: 56,
+  } satisfies CSSProperties,
+  colorSwatchLight: {
+    flex: 1,
+    boxShadow: 'inset 0 0 0 8px rgba(255, 255, 255, 0.9)',
+  } satisfies CSSProperties,
+  colorSwatchDark: {
+    flex: 1,
+    boxShadow: 'inset 0 0 0 8px rgba(17, 17, 17, 0.9)',
+  } satisfies CSSProperties,
+  breakdownSection: {
+    fontFamily: 'ui-monospace, SFMono-Regular, Menlo, monospace',
+    fontSize: 12,
+    display: 'grid',
+    gridTemplateColumns: 'auto 1fr',
+    columnGap: 12,
+    rowGap: 3,
+    marginTop: 6,
+  } satisfies CSSProperties,
+  breakdownKey: {
+    color: 'var(--sb-color-sys-text-muted, CanvasText)',
+  } satisfies CSSProperties,
+  breakdownLayerHeader: {
+    gridColumn: '1 / -1',
+    fontSize: 10,
+    textTransform: 'uppercase',
+    letterSpacing: 0.5,
+    color: 'var(--sb-color-sys-text-muted, CanvasText)',
+    marginTop: 4,
+  } satisfies CSSProperties,
+  fontFamilySample: {
+    padding: '4px 0',
+    fontSize: 22,
+    lineHeight: 1.2,
+  } satisfies CSSProperties,
+  fontWeightSample: {
+    padding: '4px 0',
+    fontSize: 32,
+    lineHeight: 1,
+  } satisfies CSSProperties,
+  dimensionTrack: {
+    display: 'flex',
+    alignItems: 'center',
+    height: 32,
+    maxWidth: '100%',
+    overflow: 'hidden',
+  } satisfies CSSProperties,
+  dimensionBar: {
+    height: 16,
+    background: 'var(--sb-color-sys-accent-bg, #3b82f6)',
+    borderRadius: 3,
+    maxWidth: '100%',
+  } satisfies CSSProperties,
+  motionTrack: {
+    position: 'relative',
+    height: 32,
+    width: '100%',
+    maxWidth: 320,
+    background: 'var(--sb-color-sys-surface-muted, rgba(128,128,128,0.08))',
+    borderRadius: 16,
+    overflow: 'hidden',
+  } satisfies CSSProperties,
+  motionBall: {
+    position: 'absolute',
+    top: '50%',
+    width: 24,
+    height: 24,
+    marginTop: -12,
+    borderRadius: '50%',
+    background: 'var(--sb-color-sys-accent-bg, #3b82f6)',
+  } satisfies CSSProperties,
+  aliasedByList: {
+    listStyle: 'none',
+    margin: 0,
+    padding: 0,
+    fontFamily: 'ui-monospace, SFMono-Regular, Menlo, monospace',
+    fontSize: 12,
+  } satisfies CSSProperties,
+  aliasedByRow: {
+    padding: '2px 0',
+    display: 'flex',
+    alignItems: 'center',
+    gap: 6,
+  } satisfies CSSProperties,
+  aliasedByTruncated: {
+    fontSize: 11,
+    opacity: 0.6,
+    fontStyle: 'italic',
+    marginTop: 4,
+  } satisfies CSSProperties,
+  reducedMotion: {
+    fontSize: 11,
+    color: 'var(--sb-color-sys-text-muted, CanvasText)',
+    fontStyle: 'italic',
+  } satisfies CSSProperties,
+};
